refactor(combo.common): update common API
